chore: move backup/restore to own module
